requirements.txt
amqp==5.0.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
certifi==2019.9.11
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-crontab==0.7.1
djangorestframework==3.12.2
fake-useragent==0.1.11
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
lxml==4.6.2
pipenv==2018.11.26
prompt-toolkit==3.0.14
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.2.5
zipp==3.4.0

pip~=21.0
setuptools~=52.0.0
django-filter~=2.4.0